+++ src/Player.h
@@ -8,7 +8,7 @@ class Player : public MovingObject
 {
 public:
     /// @param speed pixels per second
-    Player(int x_position, int y_position, double speed);
+    Player(uint16_t x_position, uint16_t y_position, double speed);
 
     /// @param delta_s time since last frame in seconds
     void update(double delta) override;
@@ -25,7 +25,7 @@ public:
     /// @param display display to draw on
     /// @param x_position X-position of drawing to clear
     /// @param y_position Y-position of the drawing to clear
-    void undraw(Display display, int x_position, int y_position) override;
+    void undraw(Display display, uint16_t x_position, uint16_t y_position) override;
 
 private:
     /// @brief in radians per second (so small numbers)
